Update seeing_a_delay_with_your_first_few_sends.md
"The length of this evaluation period can vary but you can always [contact support](https://support.sendgrid.com) to see about expediting the process." is slightly misleading as we in Support are actually not able to expedite this in any way. We can flush queues, but if they are getting longer EE times due to this, we can't speed it up for them.